Correction for clustered sandwich estimator
We’ve identified an issue with the code we sent, where the code didn’t correctly specify the clustered sandwich estimator. As a result, we need to ask analysts involved in the WFC trios GWAS to re-run two population-based models both with and without adjusting for PCs. The scripts to perform the re-runs are 3.0_unified_regression_ldak_FIX.sh and 3.1_unified_regression_ldak_sex_FIX.sh.